line count, skip fetched files

+++ getFile.py
@@ -1,12 +1,19 @@
 import csv
 import requests
+import os.path
  
-with open('sample.csv', newline='') as csvfile:
+with open('comps.csv', newline='') as csvfile:
     reader = csv.reader(csvfile, delimiter=',')
+    i = 0
     for line in reader:
-        saveas = '-'.join([line[0], line[1].replace('/', '-'), line[2].replace('/', '-'), line[3]])
+        i += 1
+        print ('i: ', i)
+        saveas = '-'.join([line[0], line[1].replace('/', '-'), line[2].replace('/', '-'), line[3]]) + '.txt'
         # Reorganize to rename the output filename.
         url = 'https://www.sec.gov/Archives/' + line[4].strip()
-        with open(saveas, 'wb') as f:
-            f.write(requests.get('%s' % url).content)
-            print(url, 'downloaded and wrote to text file')
+        if not os.path.isfile(saveas):
+            with open(saveas, 'wb') as f:
+                f.write(requests.get('%s' % url).content)
+                print(url, 'downloaded and wrote to text file')
+        else:
+            print('skipping: ',  saveas)
